Update README.md
This program is a simple contract written in Solidity, a programming language used for developing smart contracts on the Ethereum blockchain. The contract has two main functions that update the state of the contract called mint and burn. Any address can mint, but to burn, you need to burn from an address with enough funds. Feel free to play around.